bulkCraft™ - tnt/firework shulker crafting (holy win) - some fixes

+++ src/main/kotlin/co/akoot/plugins/plushies/commands/ItemEditCommand.kt
@@ -123,7 +123,7 @@ class ItemEditCommand(plugin: FoxPlugin) : FoxCommand(plugin, "edititem") {
 
                     else -> {
                         ItemBuilder.builder(item)
-                            .customModelData(cmd.toInt())
+                            .customModelData(cmd)
                             .build()
                     }
                 }
